requirements.txt
flask
flask-sqlalchemy
openai
python-dotenv